requirements.txt
django
zope.interface
mssql-django